+++ etl_plugins/utils/__init__.py
@@ -1,0 +1,19 @@
+"""Utilities: retry, chunking, async IO bridges."""
+
+from etl_plugins.utils.async_io import (
+    gather_with_concurrency,
+    iter_to_async,
+    run_sync_in_thread,
+)
+from etl_plugins.utils.chunk import chunked, drop, take
+from etl_plugins.utils.retry import retryable
+
+__all__ = [
+    "chunked",
+    "drop",
+    "gather_with_concurrency",
+    "iter_to_async",
+    "retryable",
+    "run_sync_in_thread",
+    "take",
+]